Card Drawing Working!
The card drawing is finished! It works by saving the players current cards in the database then when drawing a card it compares the new card's cardPack and cardID. If no one else has the card, it is given.

Next up is the card playing function. When you click on your card it already saves it to the database and the game already detects that the player has played said card, however the game doesn't do anything with this info. So I just have to create the card in the div and detect when everyone has played, then have the cardCzar choose one!

I thought that the card drawing function would be the hardest, so the hard part is (hopefully) over :)